renamed a couple of files that had php islands into actual php files
